+++ plugins/sdr_sources/hackrf_sdr_support/hackrf_sdr.cpp
@@ -138,7 +138,7 @@ void HackRFSource::drawControlUI()
 
 void HackRFSource::set_samplerate(uint64_t samplerate)
 {
-    for (int i = 0; i < available_samplerates.size(); i++)
+    for (int i = 0; i < (int)available_samplerates.size(); i++)
     {
         if (samplerate == available_samplerates[i])
         {
