+++ plugins/sdr_sources/spyserver_support/spyserver.cpp
@@ -185,7 +185,7 @@ void SpyServerSource::drawControlUI()
 
 void SpyServerSource::set_samplerate(uint64_t samplerate)
 {
-    for (int i = 0; i < available_samplerates.size(); i++)
+    for (int i = 0; i < (int)available_samplerates.size(); i++)
     {
         if (samplerate == available_samplerates[i])
         {
