+++ src-core/common/image/brightness_contrast.cpp
@@ -11,7 +11,7 @@ namespace image
         float brightness_v = brightness / 2.0f;
         float slant = tanf((contrast + 1.0f) * 0.78539816339744830961566084581987572104929234984378f);
 
-        for (int i = 0; i < image.height() * image.width() * channelCount; i++)
+        for (size_t i = 0; i < image.height() * image.width() * channelCount; i++)
         {
             float v = float(image[i]) / scale;
